Split the populator image
This patch will split the populator image.
One image acts as a controller to the populator - 'populator-controller'
and the actual populator action will be executed by the
'ovirt-populator' image.

This will allow a separation in logic, allowing to add more source
providers images and use the same controller for all.
