+++ hack/ovirt-populator/Containerfile
@@ -1,4 +1,4 @@
-FROM golang:1.19.3-alpine3.16 as builder
+FROM golang:1.19.5-alpine3.17 as builder
 WORKDIR /app
 
 COPY . ./
@@ -6,7 +6,6 @@ WORKDIR /app/cmd/ovirt-populator
 
 # When debugging by replacing the lib-volume-populator with a local modified copy
 # COPY . ./
-
 RUN CGO_ENABLED=0 go build -o /ovirt-populator
 
 FROM quay.io/centos/centos:stream9-minimal
